requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
faster-whisper
python-telegram-bot
requests
watchdog
pydub
python-multipart
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib